update_jquants_v1/requirements.txt
pandas
gspread
oauth2client
aiohttp
google-cloud-storage
google-cloud-secret-manager
requests
tomli
flask